+++ R/s_attribute.R
@@ -1,54 +1,61 @@
 #' Read, process and write data on structural attributes.
 #' 
-#' @details In addition to using CWB functionality, the \code{s_attribute_encode}
-#' function includes a pure R implementation to add or modify structural attributes
-#' of an existing CWB corpus.
+#' @details `s_attribute_encode()` implements a 'pure R' implementation to add
+#' or modify structural attributes of an existing CWB corpus.
 #' 
-#' If the corpus has been loaded/used before,
-#' a new s-attribute may not be available unless \code{RcppCWB::cl_delete_corpus}
-#' has been called. Use the argument \code{delete} for calling this function.
-#' @param values A character vector with the values of the structural attribute.
+#' If the corpus has been loaded/used before, a new s-attribute may not be
+#' available unless `RcppCWB::cl_delete_corpus()` has been called. Use the
+#' argument `delete` for calling this function.
+#' @param values A `character` vector with the values of the structural
+#'   attribute.
 #' @param data_dir The data directory where to write the files.
-#' @param s_attribute Atomic character vector, the name of the structural attribute.
-#' @param region_matrix A two-column \code{matrix} with corpus positions.
+#' @param s_attribute Atomic character vector, name of the structural attribute.
+#' @param region_matrix A two-column `matrix` with corpus positions.
 #' @param corpus A CWB corpus.
-#' @param method EWither 'R' or 'CWB'.
+#' @param method Either 'R' or 'CWB'.
 #' @param encoding Encoding of the data.
 #' @param registry_dir Path name of the registry directory.
-#' @param delete Logical, whether a call to \code{RcppCWB::cl_delete_corpus} is performed.
+#' @param delete Logical, whether to call `RcppCWB::cl_delete_corpus()`.
 #' @param verbose Logical.
 #' @export s_attribute_encode
 #' @importFrom RcppCWB cl_delete_corpus
-#' @seealso To decode a structural attribute, see \code{\link[RcppCWB]{s_attribute_decode}}.
+#' @seealso To decode a structural attribute, see
+#'   \code{\link[RcppCWB]{s_attribute_decode}}.
 #' @examples
 #' require("RcppCWB")
 #' registry_tmp <- fs::path(tempdir(), "cwb", "registry")
 #' data_dir_tmp <- fs::path(tempdir(), "cwb", "indexed_corpora", "reuters")
 #' 
+#' cwb_dir_rcppcwb <- system.file(package = "RcppCWB", "extdata", "cwb")
+#' registry_dir_rcppcwb <- fs::path(cwb_dir_rcppcwb, "registry")
+#' data_dir_rcppcwb <- fs::path(cwb_dir_rcppcwb,"indexed_corpora", "reuters")
+#' 
 #' corpus_copy(
 #'   corpus = "REUTERS",
-#'   registry_dir = system.file(package = "RcppCWB", "extdata", "cwb", "registry"),
-#'   data_dir = system.file(package = "RcppCWB", "extdata", "cwb", "indexed_corpora", "reuters"),
+#'   registry_dir = registry_dir_rcppcwb,
+#'   data_dir = data_dir_rcppcwb,
 #'   registry_dir_new = registry_tmp,
 #'   data_dir_new = data_dir_tmp
 #' )
 #' 
 #' no_strucs <- cl_attribute_size(
 #'   corpus = "REUTERS",
-#'   attribute = "id", attribute_type = "s",
+#'   attribute = "id",
+#'   attribute_type = "s",
 #'   registry = registry_tmp
 #' )
-#' cpos_list <- lapply(
-#'   0L:(no_strucs - 1L),
-#'   function(i)
-#'     cl_struc2cpos(corpus = "REUTERS", struc = i, s_attribute = "id", registry = registry_tmp)
+#' 
+#' cpos_matrix <- get_region_matrix(
+#'       corpus = "REUTERS",
+#'       struc = 0L:(no_strucs - 1L),
+#'       s_attribute = "id",
+#'       registry = registry_tmp
 #' )
-#' cpos_matrix <- do.call(rbind, cpos_list)
 #' 
 #' s_attribute_encode(
-#'   values = as.character(1L:nrow(cpos_matrix)),
+#'   values = 1L:nrow(cpos_matrix),
 #'   data_dir = data_dir_tmp,
-#'   s_attribute = "foo",
+#'   s_attribute = "id",
 #'   corpus = "REUTERS",
 #'   region_matrix = cpos_matrix,
 #'   method = "R",
@@ -59,13 +66,31 @@
 #' )
 #' 
 #' cl_struc2str(
-#'   "REUTERS", struc = 0L:(nrow(cpos_matrix) - 1L), s_attribute = "foo", registry = registry_tmp
+#'   "REUTERS",
+#'   struc = 0L:(nrow(cpos_matrix) - 1L),
+#'   s_attribute = "id",
+#'   registry = registry_tmp
 #' )
 #' 
 #' unlink(registry_tmp, recursive = TRUE)
 #' unlink(data_dir_tmp, recursive = TRUE)
 #' @rdname s_attribute
 s_attribute_encode <- function(values, data_dir, s_attribute, corpus, region_matrix, method = c("R", "CWB"), registry_dir = Sys.getenv("CORPUS_REGISTRY"), encoding, delete = FALSE, verbose = TRUE){
+  
+  if (!is.character(values)){
+    if (verbose)
+      cli_alert_warning(
+        "class of input {.arg values} is {.val {typeof(values)}}"
+      )
+    values <- as.character(values)
+    values_n <- length(unique(values))
+    if (verbose){
+      cli_alert_info(
+        "unique values after coercion to `character` vector: {.val {values_n}}"
+      )
+    }
+  }
+  
   stopifnot(
     inherits(region_matrix, "matrix"),
     ncol(region_matrix) == 2L,
@@ -76,19 +101,32 @@ s_attribute_encode <- function(values, data_dir, s_attribute, corpus, region_mat
     length(verbose) == 1L,
     is.logical(verbose)
   )
+  
   if (!inherits(as.vector(region_matrix), "integer")){
-    region_matrix <- matrix(data = as.integer(as.vector(region_matrix)), ncol = 2)
+    region_matrix <- matrix(
+      data = as.integer(as.vector(region_matrix)),
+      ncol = 2
+    )
   }
+  
   if (method == "R"){
-    avs_file <- fs::path(data_dir, paste(s_attribute, "avs", sep = ".")) # attribute values
-    avx_file <- fs::path(data_dir, paste(s_attribute, "avx", sep = ".")) # attribute value index
-    rng_file <- fs::path(data_dir, paste(s_attribute, "rng", sep = ".")) # ranges
+    # attribute values
+    avs_file <- fs::path(data_dir, paste(s_attribute, "avs", sep = "."))
+    # attribute value index
+    avx_file <- fs::path(data_dir, paste(s_attribute, "avx", sep = "."))
+    # ranges
+    rng_file <- fs::path(data_dir, paste(s_attribute, "rng", sep = "."))
     
     # generate and write attrib.avs
     if (!is.character(values)) values <- as.character(values)
     values_unique <- unique(values)
     if (encoding == "latin1"){
-      values_hex_list <- iconv(x = values_unique, from = "UTF-8", to = toupper(encoding), toRaw = TRUE)
+      values_hex_list <- iconv(
+        x = values_unique,
+        from = "UTF-8",
+        to = toupper(encoding),
+        toRaw = TRUE
+      )
     } else {
       values_hex_list <- lapply(values_unique, charToRaw)
     }
@@ -117,14 +155,29 @@ s_attribute_encode <- function(values, data_dir, s_attribute, corpus, region_mat
     # adjust encoding, if necessary
     input_enc <- get_encoding(as.character(tab[["s_attribute"]]))
     if (input_enc != encoding){
-      tab[["s_attribute"]] <- iconv(tab[["s_attribute"]], from = input_enc, to = encoding)
+      tab[["s_attribute"]] <- iconv(
+        tab[["s_attribute"]],
+        from = input_enc,
+        to = encoding
+      )
       Encoding(tab[["s_attribute"]]) <- encoding
     }
     
     tmp_file <- tempfile()
-    data.table::fwrite(x = tab, file = tmp_file, quote = FALSE, sep = "\t", col.names = FALSE)
+    data.table::fwrite(
+      x = tab,
+      file = tmp_file,
+      quote = FALSE,
+      sep = "\t",
+      col.names = FALSE
+    )
     
-    if (verbose) message(sprintf("... running 'cwb-s-encode' to add structural annotation for attribute '%s'", s_attribute))
+    if (verbose){
+      cli_alert_info(
+        "run 'cwb-s-encode' to add structural attribute {.val {s_attribute}}"
+      )
+    }
+    
     cmd <- c(
       fs::path(cwb_get_bindir(), "cwb-s-encode"),
       "-d", data_dir,
@@ -135,13 +188,23 @@ s_attribute_encode <- function(values, data_dir, s_attribute, corpus, region_mat
     system(paste(cmd, collapse = " "))
     
   }
-  regdata <- registry_file_parse(tolower(corpus), registry_dir = registry_dir)
+  regdata <- registry_file_parse(
+    tolower(corpus),
+    registry_dir = registry_dir
+  )
+  
   if (!s_attribute %in% regdata[["s_attributes"]]){
-    if (verbose) message(sprintf("... adding s-attribute '%s' to registry", s_attribute))
+    if (verbose)
+      cli_alert_info("add s-attribute {.val {s_attribute}} to registry")
     regdata[["s_attributes"]] <- c(regdata[["s_attributes"]], s_attribute)
-    registry_file_write(regdata, corpus = tolower(corpus), registry_dir = registry_dir)
+    registry_file_write(
+      regdata,
+      corpus = tolower(corpus),
+      registry_dir = registry_dir
+    )
   }
-  if (delete) cl_delete_corpus(corpus = toupper(corpus), registry = registry_dir)
+  if (delete)
+    cl_delete_corpus(corpus = toupper(corpus), registry = registry_dir)
   invisible( NULL )
 }